abstract code for organize and reuse
